rename SnappyWriter
