+++ src/main/java/com/spotify/sparkey/CompressionType.java
@@ -52,7 +52,7 @@ public enum CompressionType {
 
     @Override
     BlockOutput createBlockOutput(FileDescriptor fd, OutputStream outputStream, int maxBlockSize, int maxEntriesPerBlock) throws IOException {
-      return new SnappyWriter(new CompressedOutputStream(CompressorType.SNAPPY, maxBlockSize, outputStream, fd), maxEntriesPerBlock);
+      return new CompressedWriter(new CompressedOutputStream(CompressorType.SNAPPY, maxBlockSize, outputStream, fd), maxEntriesPerBlock);
     }
   },;
